move: 댓글 api 함수 폴더 구조 변경 (#205)
