+++ src/app/(mypage)/comments/page.tsx
@@ -1,8 +1,8 @@
 import { fetchGetCurrentUser } from '@/lib/apis/auth/auth-server.api';
-import { fetchAllCommentsByUserId } from '@/lib/apis/comments/get-user-comments.api';
 
 import MypageDataCounts from '../_components/mypage-data-counts';
 import CommentsList from './_components/comments-list';
+import { fetchAllCommentsByUserId } from '@/lib/apis/comments/server-comments.api';
 
 const PAGE_SIZE = 10;
